+++ src/etoolbox/utils/misc.py
@@ -1,6 +1,8 @@
 """Miscellaneous helpers and utilities."""
 
 import http.client as httplib
+import logging
+from contextlib import contextmanager
 from io import DEFAULT_BUFFER_SIZE, BytesIO
 from pathlib import Path
 
@@ -88,3 +90,31 @@ def have_internet(host: str = "8.8.8.8") -> bool:
         return False
     finally:
         conn.close()
+
+
+@contextmanager
+def all_logging_disabled(highest_level=logging.CRITICAL):
+    """Context manager to disable logging.
+
+    A context manager that will prevent any logging messages
+    triggered during the body from being processed.
+
+    Args:
+        highest_level: the maximum logging level in use.
+
+    This would only need to be changed if a custom level greater than CRITICAL
+    is defined.
+
+    two kind-of hacks here:
+        * can't get the highest logging level in effect => delegate to the user
+        * can't get the current module-level override => use an undocumented
+          (but non-private!) interface
+    """
+    previous_level = logging.root.manager.disable
+
+    logging.disable(highest_level)
+
+    try:
+        yield
+    finally:
+        logging.disable(previous_level)